backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pypdf
requests
python-dotenv